# port/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tableplan API
  version: 1.0.0-spring
  description: Checked-in contract for the Spring Boot Tableplan runtime.
servers:
  - url: /
security:
  - sessionCookie: []
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Tableplan API key with endpoint-specific scopes.
    sessionCookie:
      type: apiKey
      in: cookie
      name: TABLEPLAN_SESSION
    publicShareCookie:
      type: apiKey
      in: cookie
      name: TABLEPLAN_SHARE
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 200 }
    RecipeId:
      name: recipeId
      in: path
      required: true
      schema: { type: string, maxLength: 200 }
...
  schemas:
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string, format: uuid }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
    Session:
      type: [object, "null"]
      properties:
        user:
          type: object
          required: [id, name, email, username]
          properties:
            id: { type: string }
            name: { type: string }
            email: { type: string, format: email }
            username: { type: string }
        householdId: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication is required.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Principal, role, scope, CSRF, or Origin is denied.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, username, password]
              properties:
                name: { type: string, maxLength: 100 }
                email: { type: string, format: email, maxLength: 254 }
                username: { type: string, minLength: 3, maxLength: 32 }
                password: { type: string, minLength: 12, maxLength: 200, writeOnly: true }
      responses:
        "200": { description: Registered session., content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Account conflict. }
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string }
                password: { type: string, writeOnly: true }
      responses:
        "200": { description: Authenticated session., content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { description: Password migration is required. }
...
  /api/auth/switch-household:
    post:
      tags: [auth, household]
      operationId: switchHousehold
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [householdId], properties: { householdId: { type: string } } } } }
      responses: { "200": { description: Session switched. }, "403": { $ref: "#/components/responses/Forbidden" } }
...
  /api/v1/recipes/{recipeId}/favourite:
    get:
      tags: [recipes]
      operationId: getFavouriteState
      parameters: [{ $ref: "#/components/parameters/RecipeId" }]
      responses: { "200": { description: Current favourite state. } }
    put:
      tags: [recipes]
      operationId: setFavourite
      parameters: [{ $ref: "#/components/parameters/RecipeId" }]
      responses: { "200": { description: Favourite state updated. } }
...
  /api/v1/saved-searches:
    get:
      tags: [recipes]
      operationId: listSavedSearches
      responses: { "200": { description: Saved searches. } }
    post:
      tags: [recipes]
      operationId: saveRecipeSearch
      responses: { "200": { description: Search saved. }, "400": { $ref: "#/components/responses/BadRequest" } }
  /api/v1/saved-searches/{id}:
    delete:
      tags: [recipes]
      operationId: deleteSavedSearch
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: Search deleted. } }
...